webkit-efl/LayoutTests/platform/efl-slp/media/media-fullscreen-not-in-document-expected.txt

CONSOLE MESSAGE: line 99: TypeError: 'undefined' is not a function (evaluating 'mediaElement.load()')
FAIL: Timed out waiting for notifyDone to be called
Test media element fullscreen API when an element is not in the DOM.




*** Creating <video> element with "content/counting.mp4" not in the document, should support fullscreen because it has video
